Refatoração mapeamento da resposta da simulação.
