+++ app/src/main/java/br/com/easynvest/calc/data/repository/ResponseMapper.kt
@@ -6,11 +6,10 @@ import br.com.easynvest.calc.entity.BaseResult
 import br.com.easynvest.calc.entity.ResultBody
 import br.com.easynvest.calc.entity.ResultFooter
 import br.com.easynvest.calc.entity.ResultHeader
+import br.com.easynvest.calc.ext.brazilianMoneyFormat
+import br.com.easynvest.calc.ext.dateFormatDDMMYYYY
+import br.com.easynvest.calc.ext.formatNumberAndAddPercentSignal
 import io.reactivex.Single
-import java.text.NumberFormat
-import java.text.ParseException
-import java.text.SimpleDateFormat
-import java.util.Locale
 
 class ResponseMapper {
 
@@ -30,32 +29,26 @@ class ResponseMapper {
         val list = mutableListOf<ResultBody>()
 
         val investedAmount =
-            brazilianMoneyFormat(simulateResponse.investmentParameter.investedAmount)
+            simulateResponse.investmentParameter.investedAmount.brazilianMoneyFormat()
         list.add(ResultBody(InvestmentField.INVESTEDAMOUNT.name, investedAmount))
 
         val grossAmount =
-            brazilianMoneyFormat(simulateResponse.grossAmount)
+            simulateResponse.grossAmount.brazilianMoneyFormat()
         list.add(ResultBody(InvestmentField.GROSSAMOUNT.name, grossAmount))
 
         val grossAmountProfit =
-            brazilianMoneyFormat(simulateResponse.grossAmountProfit)
+            simulateResponse.grossAmountProfit.brazilianMoneyFormat()
         list.add(ResultBody(InvestmentField.GROSSAMOUNTPROFIT.name, grossAmountProfit))
 
         val taxesAmount =
-            brazilianMoneyFormat(simulateResponse.taxesAmount)
+            simulateResponse.taxesAmount.brazilianMoneyFormat()
         list.add(ResultBody(InvestmentField.TAXESAMOUNT.name, taxesAmount))
 
-        val netAmountProfit =
-            brazilianMoneyFormat(simulateResponse.netAmountProfit)
-        val taxesRate =
-            percentNumberFormat(simulateResponse.taxesRate)
-        list.add(ResultBody(InvestmentField.NETAMOUNT.name, netAmountProfit + taxesRate))
+        val netAmountProfit = simulateResponse.netAmountProfit.brazilianMoneyFormat()
+        val taxesRate = simulateResponse.taxesRate.formatNumberAndAddPercentSignal()
+        list.add(ResultBody(InvestmentField.NETAMOUNT.name, "$netAmountProfit[$taxesRate]"))
 
-        val maturityDate = parseDate(
-            simulateResponse.investmentParameter.maturityDate,
-            SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss", Locale.US),
-            SimpleDateFormat("dd/MM/yyyy", Locale("pt", "BR"))
-        )
+        val maturityDate = simulateResponse.investmentParameter.maturityDate.dateFormatDDMMYYYY()
         if (maturityDate == null) throw RuntimeException("Error on parse date $maturityDate")
 
         list.add(ResultBody(InvestmentField.MATURITYDATE.name, maturityDate))
@@ -65,49 +58,27 @@ class ResponseMapper {
         list.add(ResultBody(InvestmentField.MATURITYTOTALDAYS.name, maturityTotalDays))
 
         val monthlyGrossRateProfit =
-            percentNumberFormat(simulateResponse.annualGrossRateProfit)
-        list.add(
-            ResultBody(
-                InvestmentField.MONTHLYGROSSRATEPROFIT.name,
-                monthlyGrossRateProfit
-            )
-        )
+            simulateResponse.annualGrossRateProfit.formatNumberAndAddPercentSignal()
+        list.add(ResultBody(InvestmentField.MONTHLYGROSSRATEPROFIT.name, monthlyGrossRateProfit))
 
-        val rate =
-            percentNumberFormat(simulateResponse.investmentParameter.rate)
+        val rate = simulateResponse.investmentParameter.rate.formatNumberAndAddPercentSignal()
         list.add(ResultBody(InvestmentField.RATE.name, rate))
 
         val annualGrossRateProfit =
-            percentNumberFormat(simulateResponse.annualGrossRateProfit)
+            simulateResponse.annualGrossRateProfit.formatNumberAndAddPercentSignal()
         list.add(ResultBody(InvestmentField.ANNUALGROSSRATEPROFIT.name, annualGrossRateProfit))
 
-        val rateProfit =
-            percentNumberFormat(simulateResponse.rateProfit)
+        val rateProfit = simulateResponse.rateProfit.formatNumberAndAddPercentSignal()
         list.add(ResultBody(InvestmentField.RATEPROFIT.name, rateProfit))
 
         return list
     }
 
     private fun mapHeader(simulateResponse: SimulateResponse): ResultHeader {
-        val grossAmount = brazilianMoneyFormat(simulateResponse.grossAmount)
-        val annualGrossRateProfit = brazilianMoneyFormat(simulateResponse.annualGrossRateProfit)
+        val grossAmount = simulateResponse.grossAmount.brazilianMoneyFormat()
+        val annualGrossRateProfit = simulateResponse.annualGrossRateProfit.brazilianMoneyFormat()
 
         return ResultHeader(grossAmount, annualGrossRateProfit)
     }
-
-    private fun brazilianMoneyFormat(parsed: Double): String =
-        NumberFormat.getCurrencyInstance(Locale("pt", "BR")).format(parsed)
-
-    private fun parseDate(
-        inputDateString: String, input: SimpleDateFormat, output: SimpleDateFormat
-    ): String? =
-        try {
-            val parse = input.parse(inputDateString)
-            if (parse != null) output.format(parse) else throw  RuntimeException("Input date format error")
-        } catch (e: ParseException) {
-            throw RuntimeException("Error on parse date")
-        }
-
-    private fun percentNumberFormat(parsed: Double): String = "$parsed%"
 }
